0-print_list.c

+++ singly_linked_lists/0-print_list.c
@@ -18,14 +18,10 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%u] %s\n", (unsigned long)h->len, h->str);
 		}
 	h = h->next;
 	i++;
 	}
 	return (i);
 }
-
-
-
-
